Chunk the engine's log messages instead of truncating them
Dart print() and debugPrint() reach the shell through
Engine::onLogMessageCallback, which formatted "{}: {}" into one record
and let the rest fall off. A record holds IHS_LOG_TEXT_CAPACITY bytes,
so with the "flutter" tag a message longer than 230 bytes lost its tail
-- and a stack trace or an encoded summary keeps the part worth reading
at the end.

The shell already decided how to handle this. LoggingHandler splits long
messages on newlines and hard-splits the remainder, with a comment
saying it does so "so nothing is silently truncated". That treatment
just never reached the other path Dart output takes.

Lift that helper into shell/logging/log_chunk.h, give it an explicit
budget so a caller that prefixes each record can subtract its prefix,
and use it from both places. The engine callback repeats the tag on
every chunk, so it passes the capacity minus "<tag>: ".

Verified on a Raspberry Pi 4 on drm-kms-egl, using multi_touch_test with
SELFCHECK_AFTER_S so the summary prints on a timer and no input is
needed. Its JSON is 301 bytes. Before, the line ended at 230 and the
next line was the following summary -- the remainder never existed.
After, the 230-byte line is followed by a 71-byte continuation carrying
the rest. Re-ran on the post-clang-format binary.
